Fix install script to work with new Database API
Implementation of the new database API has broken the install script.
This patch updates the install script to work with the current state of
the database API. In particular, we now use exceptions for getting
errors from the database.

Also fixed an XHTML strictness bug with a closing </tr> being
out-of-place.